+++ src/i18n/strings/en_EN.json
@@ -57,6 +57,7 @@
     "Server may be unavailable, overloaded, or you hit a bug.": "Server may be unavailable, overloaded, or you hit a bug.",
     "The server does not support the room version specified.": "The server does not support the room version specified.",
     "Failure to create room": "Failure to create room",
+    "Send cross-signing keys to homeserver": "Send cross-signing keys to homeserver",
     "Send anyway": "Send anyway",
     "Send": "Send",
     "Sun": "Sun",
@@ -495,7 +496,6 @@
     "New Password": "New Password",
     "Confirm password": "Confirm password",
     "Change Password": "Change Password",
-    "Send cross-signing keys to homeserver": "Send cross-signing keys to homeserver",
     "Cross-signing public keys:": "Cross-signing public keys:",
     "on device": "on device",
     "not found": "not found",
